Move session expiry code into a new class
