+++ src/main/authentication/authenticationHandler.ts
@@ -1,5 +1,3 @@
-import moment from 'moment';
-
 import process from 'process';
 import config from 'config';
 import {AccountManagementRequests} from '../resources/requests/accountManagementRequests';
@@ -12,26 +10,6 @@ export const mediaAccountCreationRoles = ['INTERNAL_SUPER_ADMIN_CTSC', 'INTERNAL
 export const allAdminRoles = ['SYSTEM_ADMIN', 'INTERNAL_SUPER_ADMIN_CTSC', 'INTERNAL_SUPER_ADMIN_LOCAL', 'INTERNAL_ADMIN_CTSC', 'INTERNAL_ADMIN_LOCAL'];
 export const verifiedRoles = ['VERIFIED'];
 const FRONTEND_URL = process.env.FRONTEND_URL || 'https://pip-frontend.staging.platform.hmcts.net';
-
-export function isSessionExpired(req): boolean {
-  if (req.user === undefined) {
-    return false;
-  }
-
-  if(req.session.sessionExpires) {
-    const sessionExpiryDateTime = moment.utc(req.session.sessionExpires);
-    const currentDateTime = moment.utc(new Date(Date.now()));
-    const durationAsSeconds = moment.duration(sessionExpiryDateTime.startOf('seconds').diff(currentDateTime.startOf('seconds'))).asSeconds();
-    if(durationAsSeconds <= 0) {
-      return true;
-    }
-  }
-
-  const defaultSessionExpiry = 60 * 60 * 1000; // default to 1 hour
-  const sessionExpiry = checkRoles(req, verifiedRoles) ? defaultSessionExpiry : 4 * defaultSessionExpiry;
-  req.session.sessionExpires = new Date(Date.now() + sessionExpiry);
-  return false;
-}
 
 export function checkRoles(req, roles): boolean {
   if(req.user) {
